plugins/tbe07tyg/pdf_reader/requirements.txt
PyPDF2>=3.0.0
requests>=2.28.0
python-magic>=0.4.27 